DjangoShellyElectiricityAutomation/requirements.txt
amqp
asgiref
beautifulsoup4
billiard
bs4
certifi
charset_normalizer
click
click_didyoumean
click_plugins
click_repl
colorama
cron_descriptor
django
django-celery-beat
django-timezone-field
entsoe-py
idna
kombu
numpy
pandas
pip
prompt_toolkit
python-crontab
python-dateutil
pytz
redis
requests
six
soupsieve
sqlparse
typing-extensions
tzdata
urllib3
vine
wcwidth
django-apscheduler

